fix: Refactor component imports and fix file paths
